requirements.txt
transformers
Pillow
pypdf
jq
langchain
streamlit
boto3
awscli
pandas-datareader
langchain-experimental
bs4
setuptools
langchain_community
python-dotenv
botocore